AbstractItemResponseModels

Documentation for AbstractItemResponseModels.

AbstractItemResponseModels.Continuous

AbstractItemResponseModels.Dichotomous

AbstractItemResponseModels.Dimensionality

AbstractItemResponseModels.EstimationType

AbstractItemResponseModels.ItemResponseModel

AbstractItemResponseModels.Multivariate

AbstractItemResponseModels.Nominal

AbstractItemResponseModels.Ordinal

AbstractItemResponseModels.ResponseType

AbstractItemResponseModels.Univariate

AbstractItemResponseModels.expected_score

AbstractItemResponseModels.fit

AbstractItemResponseModels.irf

AbstractItemResponseModels.item_dimensionality

AbstractItemResponseModels.person_dimensionality

AbstractItemResponseModels.predict

AbstractItemResponseModels.response_type

AbstractItemResponseModels.Continuous — Type

Continous <: ResponseType

Defines that an ItemResponseModel has a continous response variable.

AbstractItemResponseModels.Dichotomous — Type

Dichotomous <: ResponseType

Defines that an ItemResponseModel has a binary response variable.

AbstractItemResponseModels.Dimensionality — Type

Dimensionality

The Dimensionality of an ItemResponseModel defines whether or not the person and/or item parameters of the model are unidimensional or multidimensional.

AbstractItemResponseModels.EstimationType — Type

AbstractItemResponseModels.ItemResponseModel — Type

AbstractItemResponseModels.Multivariate — Type

Multivariate <: Dimensionality

Defines that an ItemResponseModel has vector valued parameters.

AbstractItemResponseModels.Nominal — Type

Nominal <: ResponseType

Defines that an ItemResponseModel has an unordered categorical response variable.

AbstractItemResponseModels.Ordinal — Type

Ordinal <: ResponseType

Defines that an ItemResponseModel has an ordered categorical response variable.

AbstractItemResponseModels.ResponseType — Type

ResponseType

The ResponseType defines the scale level of item responses for a given ItemResponseModel.

AbstractItemResponseModels.Univariate — Type

Univariate <: Dimensionality

Defines that an ItemResponseModel has scalar valued parameters.

AbstractItemResponseModels.expected_score — Function

expected_score(model::ItemResponseModel, theta) -> Float64
expected_score(model::ItemResponseModel, theta, is) -> Float64

Calculate the expected score of an ItemResponseModel given person parameter(s) theta.

If isfitted(model) == false then the expected score can only be calculated given expilict item parameter(s) beta.

If isfitted(model) == true then expected scores can also be calculated for one or multiple items by specifying one or multiple unique item indices is. For models where item_dimensionality(model) == Univariate, is is either an Int (for a single item) or Vector{Int} (for multiple items). For models where item_dimensionality(model) == Multivariate, is is either a Vector{Int} (for a single item) or Vector{Vector{Int}} (for multiple items).

In either case expected_score returns a single value.

AbstractItemResponseModels.fit — Function

fit(::Type{<:ItemResponseModel}, data::AbstractMatrix[, alg])

Fit an item response model to response data.

If data is a response matrix, the columns must describe items and rows describe persons.

AbstractItemResponseModels.irf — Function

irf(model::ItemResponseModel, theta, i, y)

Evaluate the item response function of an ItemResponseModel given person parameter(s) theta for response y.

AbstractItemResponseModels.item_dimensionality — Method

item_dimensionality(::T) -> Dimensionality

Return the Dimensionality of the item parameters of an ItemResponseModel.

AbstractItemResponseModels.person_dimensionality — Method

item_dimensionality(::T) -> Dimensionality

Return the Dimensionality of the person parameters of an ItemResponseModel.

AbstractItemResponseModels.predict — Function

AbstractItemResponseModels.response_type — Method

response_type(::T) -> ResponseType

Return the ResponseType of the ItemResponseModel.