API

Types

AbstractItemResponseModels.ItemResponseModel — Type

ItemResponseModel

An abstract type representing an item response theory model.

Each implementation T <: ItemResponseModel must define the following traits:

response_type: A valid ResponseType

person_dimensionality: A valid Dimensionality

item_dimensionality: A valid Dimensionality

estimation_type: A valid EstimationType

Additionally T <: ItemResponseModel must implement the following interface:

irf: An item response function returning the probability that a person with given ability estimate will answer an item with a particular response.

iif: An item information function returning the information of answering with a particular response on an item given an ability estimate.

fit: A function fitting an item response model of type T to observed data.

Traits

Dimensionality

AbstractItemResponseModels.Dimensionality — Type

Dimensionality

The Dimensionality of an ItemResponseModel defines whether or not the person and/or item parameters of the model are unidimensional or multidimensional.

AbstractItemResponseModels.Univariate — Type

Univariate <: Dimensionality

Defines that an ItemResponseModel has scalar valued parameters.

AbstractItemResponseModels.Multivariate — Type

Multivariate <: Dimensionality

Defines that an ItemResponseModel has vector valued parameters.

AbstractItemResponseModels.item_dimensionality — Function

item_dimensionality(::T) -> Dimensionality

Return the Dimensionality of the item parameters of an ItemResponseModel.

AbstractItemResponseModels.person_dimensionality — Function

item_dimensionality(::T) -> Dimensionality

Return the Dimensionality of the person parameters of an ItemResponseModel.

Estimation type

AbstractItemResponseModels.EstimationType — Type

EstimationType

The EstimationType of an ItemResponseModel describes the type of parameter estimation for the model.

AbstractItemResponseModels.PointEstimate — Type

PointEstimate <: EstimationType

Defines that the parameters of an ItemResponseModel are estimated by a single point estimate. Thus, parameters for a model with estimation_type(model) == PointEstimate are single real-valued numbers.

AbstractItemResponseModels.SamplingEstimate — Type

SamplingEstimate <: EstimationType

Defines that the parameters of an ItemResponseModel are estimated by a sampling procedure. Thus, parameters for a model with estimation_type(model) == SamplingEstimate are vectors of real-valued numbers where the length of the parameter vector is equal to the number of samples drawn.

AbstractItemResponseModels.estimation_type — Function

estimation_type(::T) -> EstimationType

Return the EstimationType of an [ItemResponseModel].

Response type

AbstractItemResponseModels.ResponseType — Type

ResponseType

The ResponseType defines the scale level of item responses for a given ItemResponseModel.

AbstractItemResponseModels.Dichotomous — Type

Dichotomous <: ResponseType

Defines that an ItemResponseModel has a binary response variable.

AbstractItemResponseModels.Nominal — Type

Nominal <: ResponseType

Defines that an ItemResponseModel has an unordered categorical response variable.

AbstractItemResponseModels.Ordinal — Type

Ordinal <: ResponseType

Defines that an ItemResponseModel has an ordered categorical response variable.

AbstractItemResponseModels.Continuous — Type

Continous <: ResponseType

Defines that an ItemResponseModel has a continous response variable.

AbstractItemResponseModels.response_type — Function

response_type(::T) -> ResponseType

Return the ResponseType of an ItemResponseModel.

AbstractItemResponseModels.checkresponsetype — Function

checkresponsetype(::T, arr) where {T<:ResponseType}

Check if the responses in arr are valid for ResponseType T.

If all responses are valid, nothing is returned.

If any invalid responses are found, a DomainError is thrown.

Methods

AbstractItemResponseModels.expected_score — Function

expected_score(model::ItemResponseModel, theta; scoring_function) -> Float64
expected_score(model::ItemResponseModel, theta; scoring_function) -> Vector{Float64}
expected_score(model::ItemResponseModel, theta, is; scoring_function) -> Float64
expected_score(model::ItemResponseModel, theta, is; scoring_function) -> Vector{Float64}

Calculate the expected score of an ItemResponseModel.

Arguments

model: An ItemResponseModel

theta: The person parameter value(s)

is: One or multiple item identifiers. If is is omitted, the expected score for the whole test is returned.

Keyword arguments

scoring_function: A function mapping all possible response values y to arbitrary values. Must default to identity.

Return values

If estimation_type(model) == PointEstimate then expected_score must return a single scalar value.

If estimation_type(model) == SamplingEstimate then expected_score must return a vector of values with the length equal to the number of samples drawn.

StatsAPI.fit — Function

fit(::Type{T}, data::AbstractMatrix, args...; kwargs...) where {T<:ItemResponseModel} -> T

Fit an item response model to response data.

Arguments

data: Observed data. If data is a response matrix, the columns must describe items and rows describe persons.

args...: Additional arguments required to fit the ItemResponseModel

Keyword arguments

kwargs...: Additional keyword arguments required to fit the ItemResponseModel

Return values

A fitted ItemResponseModel.

AbstractItemResponseModels.iif — Function

iif(model::ItemResponseModel, theta, i, y) -> Float64
iif(model::ItemResponseModel, theta, i, y) -> Vector{Float64}

Argument

model: An ItemResponseModel

theta: The person parameter value(s)

i: A unique item identifier

y: Response value(s)

Return values

If estimation_type(model) == PointEstimate then the item information function must return a scalar value.

If estimation_type(model) == SamplingEstimate then the item information function must return a vector of values with the length equal to the number of samples drawn.

AbstractItemResponseModels.information — Function

information(model::ItemResponseModel, theta; scoring_function) -> Float64
information(model::ItemResponseModel, theta; scoring_function) -> Vector{Float64}
information(model::ItemResponseModel, theta, is; scoring_function) -> Float64
information(model::ItemResponseModel, theta, is; scoring_function) -> Vector{Float64}

Calculate the information of an ItemResponseModel.

Arguments

theta: The person parameter value(s)

is: One or multiple item identifiers. If is is omitted, the information of the whole test (test information) is returned.

Keyword arguments

scoring_function: A function mapping all possible response values y to arbitrary values. Must default to identity.

Return values

If estimatione_type(model) == PointEstimate then information must return a single scalar value.

If estimation_type(model) == SamplingEstimate then information musst return a vector of values with the length equal to the number of samples drawn.

AbstractItemResponseModels.irf — Function

irf(model::ItemResponseModel, theta, i, y) -> Float64
irf(model::ItemResponseModel, theta, i, y) -> Vector{Float64}

Evaluate the item response function of an ItemResponseModel.

Argument

model: An ItemResponseModel

theta: The person parameter value(s)

i: A unique item identifier

y: Response value(s)

Return values

If estimation_type(model) == PointEstimate then the item response function must return a scalar value.

If estimation_type(model) == SamplingEstimate then the item response function must return a vector of values with the length equal to the number of samples.

Tests

Interface tests are provided in the AbstractItemResponseModels.Tests module.

AbstractItemResponseModels.Tests.test_interface — Function

test_interface(T::Type{<:ItemResponseModel}, data, args...; kwargs...)

Test the implementation of an item response model interface.

Arguments

T: The implemented model type

data: observed data

args...: Additional arguments passed to fit

Keyword arguments

kwargs...: Additional keyword arguments passed to fit