AbstractItemResponseModels

Documentation for AbstractItemResponseModels.

AbstractItemResponseModels.Continuous

AbstractItemResponseModels.Dichotomous

AbstractItemResponseModels.Dimensionality

AbstractItemResponseModels.EstimationType

AbstractItemResponseModels.ItemResponseModel

AbstractItemResponseModels.Multivariate

AbstractItemResponseModels.Nominal

AbstractItemResponseModels.Ordinal

AbstractItemResponseModels.PointEstimate

AbstractItemResponseModels.ResponseType

AbstractItemResponseModels.SamplingEstimate

AbstractItemResponseModels.Univariate

AbstractItemResponseModels.checkresponsetype

AbstractItemResponseModels.estimation_type

AbstractItemResponseModels.expected_score

AbstractItemResponseModels.iif

AbstractItemResponseModels.information

AbstractItemResponseModels.irf

AbstractItemResponseModels.item_dimensionality

AbstractItemResponseModels.person_dimensionality

AbstractItemResponseModels.response_type

StatsAPI.fit

StatsAPI.predict

AbstractItemResponseModels.Continuous — Type

Continous <: ResponseType

Defines that an ItemResponseModel has a continous response variable.

AbstractItemResponseModels.Dichotomous — Type

Dichotomous <: ResponseType

Defines that an ItemResponseModel has a binary response variable.

AbstractItemResponseModels.Dimensionality — Type

Dimensionality

The Dimensionality of an ItemResponseModel defines whether or not the person and/or item parameters of the model are unidimensional or multidimensional.

AbstractItemResponseModels.EstimationType — Type

EstimationType

The EstimationType of an ItemResponseModel describes the type of parameter estimation for the model.

AbstractItemResponseModels.ItemResponseModel — Type

ItemResponseModel

An abstract type representing an item response theory model.

Each implementation T <: ItemResponseModel must define the following traits:

response_type: A valid ResponseType

person_dimensionality: A valid Dimensionality

item_dimensionality: A valid Dimensionality

estimation_type: A valid EstimationType

Additionally T <: ItemResponseModel must implement the following interface:

irf: An item response function returning the probability that a person with given ability estimate will answer an item with a particular response.

iif: An item information function returning the information of answering with a particular response on an item given an ability estimate.

fit: A function fitting an item response model of type T to observed data.

predict: A function predicting future responses

AbstractItemResponseModels.Multivariate — Type

Multivariate <: Dimensionality

Defines that an ItemResponseModel has vector valued parameters.

AbstractItemResponseModels.Nominal — Type

Nominal <: ResponseType

Defines that an ItemResponseModel has an unordered categorical response variable.

AbstractItemResponseModels.Ordinal — Type

Ordinal <: ResponseType

Defines that an ItemResponseModel has an ordered categorical response variable.

AbstractItemResponseModels.PointEstimate — Type

PointEstimate <: EstimationType

Defines that the parameters of an ItemResponseModel are estimated by a single point estimate. Thus, parameters for a model with estimation_type(model) == PointEstimate are single real-valued numbers.

AbstractItemResponseModels.ResponseType — Type

ResponseType

The ResponseType defines the scale level of item responses for a given ItemResponseModel.

AbstractItemResponseModels.SamplingEstimate — Type

SamplingEstimate <: EstimationType

Defines that the parameters of an ItemResponseModel are estimated by a sampling procedure. Thus, parameters for a model with estimation_type(model) == SamplingEstimate are vectors of real-valued numbers where the length of the parameter vector is equal to the number of samples drawn.

AbstractItemResponseModels.Univariate — Type

Univariate <: Dimensionality

Defines that an ItemResponseModel has scalar valued parameters.

AbstractItemResponseModels.checkresponsetype — Method

checkresponsetype(::T, arr) where {T<:ResponseType}

Check if the responses in arr are valid for ResponseType T.

If all responses are valid, nothing is returned.

If any invalid responses are found, a DomainError is thrown.

AbstractItemResponseModels.estimation_type — Method

estimation_type(::T) -> EstimationType

Return the EstimationType of an [ItemResponseModel].

AbstractItemResponseModels.expected_score — Function

expected_score(model::ItemResponseModel, theta) -> Float64
expected_score(model::ItemResponseModel, theta) -> Vector{Float64}
expected_score(model::ItemResponseModel, theta, is) -> Float64
expected_score(model::ItemResponseModel, theta, is) -> Vector{Float64}

Calculate the expected score of an ItemResponseModel given person parameter(s) theta and optionally one or multiple item identifiers is. If is is omitted, the expected score of the whole test is returned.

If estimation_type(model) == PointEstimate then expected_score must return a single scalar value.

If estimation_type(model) == SamplingEstimate then expected_score must return a vector of values with the length equal to the number of samples drawn.

AbstractItemResponseModels.iif — Function

iif(model::ItemResponseModel, theta, i, y) -> Float64
iif(model::ItemResponseModel, theta, i, y) -> Vector{Float64}

Evaluate the item information function of an ItemResponseModel given person parameters(s) theta for response y.

The item is specified by an unique item identifier i.

If estimation_type(model) == PointEstimate then the item information function must return a scalar value.

If estimation_type(model) == SamplingEstimate then the item information function must return a vector of values with the length equal to the number of samples drawn.

AbstractItemResponseModels.information — Function

information(model::ItemResponseModel, theta) -> Float64
information(model::ItemRepsonseModel, theta) -> Vector{Float64}
information(model::ItemResponseModel, theta, is) -> Float64
information(model::ItemRepsonseModel, theta, is) -> Vector{Float64}

Calculate the information of an ItemResponseModel given person parameter(s) theta and optionally one or multiple item identifiers is. If is is omitted, the information of the whole test (test information) is returned.

If estimatione_type(model) == PointEstimate then information must return a single scalar value.

If estimation_type(model) == SamplingEstimate then information musst return a vector of values with the length equal to the number of samples drawn.

AbstractItemResponseModels.irf — Function

irf(model::ItemResponseModel, theta, i, y) -> Float64
irf(model::ItemResponseModel, theta, i, y) -> Vector{Float64}

Evaluate the item response function of an ItemResponseModel given person parameter(s) theta for response y.

The item is specified by an unique item identitfier i.

If estimation_type(model) == PointEstimate then the item response function must return a scalar value.

If estimation_type(model) == SamplingEstimate then the item response function must return a vector of values with the length equal to the number of samples.

AbstractItemResponseModels.item_dimensionality — Method

item_dimensionality(::T) -> Dimensionality

Return the Dimensionality of the item parameters of an ItemResponseModel.

AbstractItemResponseModels.person_dimensionality — Method

item_dimensionality(::T) -> Dimensionality

Return the Dimensionality of the person parameters of an ItemResponseModel.

AbstractItemResponseModels.response_type — Method

response_type(::T) -> ResponseType

Return the ResponseType of an ItemResponseModel.

StatsAPI.fit — Function

fit(::Type{<:ItemResponseModel}, data::AbstractMatrix, args...; kwargs...)

Fit an item response model to response data.

If data is a response matrix, the columns must describe items and rows describe persons.

StatsAPI.predict — Function

predict(model::ItemResponseModel, args...; kwargs...)